+++ src/helpers/index.ts
@@ -11,10 +11,12 @@ import defineTokenSymbol from './define-token-symbol'
 import defineIfEscrowAddressIsCorrect from './define-if-escrow-address-is-correct'
 import createQueryString from './create-query-string'
 import updateOperations from './update-operations'
+import parseQueryParams from './parse-query-params'
 
 export {
   defineJSONRpcUrl,
   updateOperations,
+  parseQueryParams,
   createQueryString,
   defineIfEscrowAddressIsCorrect,
   defineTokenSymbol,