+++ src/modules/linkdrop-p2p/types/parse-version/index.ts
@@ -1,0 +1,5 @@
+type TParseVersion = (
+  claimUrl: string
+) => string
+
+export default TParseVersion